실적 / 상차 분리

+++ lib/domain/model/tb_wh_item.dart
@@ -18,7 +18,7 @@ class TbWhItem with _$TbWhItem {
     String? warehouseCd,
     int? width,
     String? useYn,
-    int? boxKg,
+    double? boxKg,
     int? height,
     int? rgstrId,
     int? updtrId,
